Bug fix: export/ directory. Feature: Confirm delete button.

+++ client/src/components/ad-request/PendingAdRequestList.vue
@@ -76,6 +76,11 @@ async function getPendingAdRequests() {
 async function deleteAdRequest(adRequestId) {
     console.log('Ad Request Deleted!');
     try {
+        const confirmDelete = window.confirm("Are you sure you want to delete this ad request?");
+        if (!confirmDelete) {
+            return null;
+        }
+        
         const res = await fetch(`/api/ad-request/${adRequestId}`, {
             method: 'DELETE',
             headers: { 'Authentication-Token': localStorage.getItem('authToken') }
