+++ client/src/views/AdvertsView.vue
@@ -101,6 +101,11 @@ async function completeAdRequest(adRequestId) {
 async function deleteAdRequest(adRequestId) {
     console.log('Ad Request Deleted!');
     try {
+        const confirmDelete = window.confirm("Are you sure you want to delete this campaign?");
+        if (!confirmDelete) {
+            return null;
+        }
+        
         const res = await fetch(`/api/ad-request/${adRequestId}`, {
             method: 'DELETE',
             headers: { 'Authentication-Token': localStorage.getItem('authToken') }
